+++ src/main/java/pl/thedeem/intellij/dql/psi/DQLElementFactory.java
@@ -1,12 +1,13 @@
 package pl.thedeem.intellij.dql.psi;
 
 import com.intellij.openapi.project.Project;
+import com.intellij.psi.PsiComment;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFile;
 import com.intellij.psi.PsiFileFactory;
 import com.intellij.psi.util.PsiTreeUtil;
-import pl.thedeem.intellij.dql.DynatraceQueryLanguage;
 import org.jetbrains.annotations.NotNull;
+import pl.thedeem.intellij.dql.DynatraceQueryLanguage;
 
 import java.util.Objects;
 
@@ -24,9 +25,16 @@ public class DQLElementFactory {
     public static PsiElement createUnknownElement(@NotNull Project project, @NotNull String text) {
         PsiFile newFile = PsiFileFactory.getInstance(project).createFileFromText("temporary.dql", DynatraceQueryLanguage.INSTANCE, "fetch " + text);
         DQLQueryStatementKeyword keyword = PsiTreeUtil.findChildOfType(newFile, DQLQueryStatementKeyword.class);
-        if (keyword == null) {
-            throw new RuntimeException("DQLElementFactory::createUnknownElement error");
-        }
-        return keyword.getNextSibling().getNextSibling();
+        return Objects.requireNonNull(keyword).getNextSibling().getNextSibling();
+    }
+
+    public static @NotNull PsiComment createInlineComment(@NotNull Project project, @NotNull String text) {
+        PsiFile newFile = PsiFileFactory.getInstance(project).createFileFromText("temporary.dql", DynatraceQueryLanguage.INSTANCE, "//" + text);
+        return Objects.requireNonNull(PsiTreeUtil.findChildOfType(newFile, PsiComment.class));
+    }
+
+    public static @NotNull PsiComment createMultiLineComment(@NotNull Project project, @NotNull String text) {
+        PsiFile newFile = PsiFileFactory.getInstance(project).createFileFromText("temporary.dql", DynatraceQueryLanguage.INSTANCE, "/*" + text + "*/");
+        return Objects.requireNonNull(PsiTreeUtil.findChildOfType(newFile, PsiComment.class));
     }
 }